fix(01): tu jsonSchema
