+++ tests/toJsonSchema/transformers/array.test.ts
@@ -1,5 +1,31 @@
 import { render, defaultTransformers } from "@scripts/toJsonSchema";
-import { DP, DPE } from "@duplojs/utils";
+import { arrayTransformer } from "@scripts/toJsonSchema/transformer/defaults";
+import {
+	supportedVersions,
+	type TransformerParams,
+} from "@scripts/toJsonSchema/transformer/create";
+import { DP, DPE, E } from "@duplojs/utils";
+
+function buildTransformerParams(
+	schema: DP.DataParser,
+	transformer: TransformerParams["transformer"],
+): TransformerParams {
+	return {
+		mode: "out",
+		context: new Map(),
+		version: supportedVersions.jsonSchema7,
+		transformer,
+		success(result, canBeUndefined = false) {
+			return E.right("buildSuccess", {
+				schema: result,
+				canBeUndefined,
+			});
+		},
+		buildError() {
+			return E.left("buildDataParserError", schema);
+		},
+	};
+}
 
 describe("array", () => {
 	it("renders items and constraints", () => {
@@ -47,5 +73,17 @@ describe("array", () => {
 				},
 			),
 		).toMatchSnapshot();
+	});
+
+	it("returns left when element transform fails", () => {
+		const schema = DPE.array(DPE.string());
+		const params = buildTransformerParams(
+			schema,
+			() => E.left("dataParserNotSupport", schema.definition.element),
+		);
+
+		expect(arrayTransformer(schema, params)).toStrictEqual(
+			E.left("dataParserNotSupport", schema.definition.element),
+		);
 	});
 });
