+++ tests/toJsonSchema/transformers/object.test.ts
@@ -1,5 +1,31 @@
 import { render, defaultTransformers } from "@scripts/toJsonSchema";
-import { DPE } from "@duplojs/utils";
+import { objectTransformer } from "@scripts/toJsonSchema/transformer/defaults";
+import {
+	supportedVersions,
+	type TransformerParams,
+} from "@scripts/toJsonSchema/transformer/create";
+import { type DP, DPE, E } from "@duplojs/utils";
+
+function buildTransformerParams(
+	schema: DP.DataParser,
+	transformer: TransformerParams["transformer"],
+): TransformerParams {
+	return {
+		mode: "out",
+		context: new Map(),
+		version: supportedVersions.jsonSchema7,
+		transformer,
+		success(result, canBeUndefined = false) {
+			return E.right("buildSuccess", {
+				schema: result,
+				canBeUndefined,
+			});
+		},
+		buildError() {
+			return E.left("buildDataParserError", schema);
+		},
+	};
+}
 
 describe("object", () => {
 	it("marks optional properties with undefined support", () => {
@@ -55,5 +81,20 @@ describe("object", () => {
 				},
 			),
 		).toMatchSnapshot();
+	});
+
+	it("returns left when a property transform fails", () => {
+		const schema = DPE.object({
+			name: DPE.string(),
+		});
+
+		const params = buildTransformerParams(
+			schema,
+			() => E.left("dataParserNotSupport", schema.definition.shape.name),
+		);
+
+		expect(objectTransformer(schema, params)).toStrictEqual(
+			E.left("dataParserNotSupport", schema.definition.shape.name),
+		);
 	});
 });
