+++ tests/toJsonSchema/transformers/tuple.test.ts
@@ -1,5 +1,31 @@
 import { render, defaultTransformers } from "@scripts/toJsonSchema";
-import { DPE } from "@duplojs/utils";
+import { tupleTransformer } from "@scripts/toJsonSchema/transformer/defaults";
+import {
+	supportedVersions,
+	type TransformerParams,
+} from "@scripts/toJsonSchema/transformer/create";
+import { DP, DPE, E } from "@duplojs/utils";
+
+function buildTransformerParams(
+	schema: DP.DataParser,
+	transformer: TransformerParams["transformer"],
+): TransformerParams {
+	return {
+		mode: "out",
+		context: new Map(),
+		version: supportedVersions.jsonSchema7,
+		transformer,
+		success(result, canBeUndefined = false) {
+			return E.right("buildSuccess", {
+				schema: result,
+				canBeUndefined,
+			});
+		},
+		buildError() {
+			return E.left("buildDataParserError", schema);
+		},
+	};
+}
 
 describe("tuple", () => {
 	it("fixed items", () => {
@@ -26,5 +52,22 @@ describe("tuple", () => {
 				},
 			),
 		).toMatchSnapshot();
+	});
+
+	it("returns left when one item transform fails", () => {
+		const schema = DPE.tuple([DPE.string(), DPE.number()]);
+		const params = buildTransformerParams(
+			schema,
+			(inner) => DP.stringKind.has(inner)
+				? E.left("dataParserNotSupport", inner)
+				: E.right("buildSuccess", {
+					schema: { type: "number" },
+					canBeUndefined: false,
+				}),
+		);
+
+		expect(tupleTransformer(schema, params)).toStrictEqual(
+			E.left("dataParserNotSupport", schema.definition.shape[0]),
+		);
 	});
 });
